+++ src/lib/commands/status.ts
@@ -1,6 +1,6 @@
 import { intervalToDuration } from 'date-fns';
 import { Client, MessageEmbed } from 'discord.js';
-import { ICommand, COMMAND_TYPE } from '../common/ICommand';
+import { ICommand, COMMAND_TYPE } from '../interfaces/ICommand';
 
 const StatusCommand: ICommand<COMMAND_TYPE.LEGACY> = {
   name: 'status',
@@ -40,7 +40,8 @@ const StatusCommand: ICommand<COMMAND_TYPE.LEGACY> = {
         {
           name: 'Info',
           value: `Repository: ${pjson.repository ?? 'unavailable'}
-      WebSocket latency: ${client.ws.ping + 'ms' ?? 'unavailable'}`,
+      WebSocket latency: ${client.ws.ping + 'ms' ?? 'unavailable'}
+      Amount of guilds joined: ${client.guilds.cache.size ?? 'unavailable'}`,
           inline: false,
         },
         {