+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.5.0](https://github.com/tatupesonen/formatbot/compare/v1.3.1...v1.5.0) (2021-09-18)
+
+
+### Features
+
+* add commitlint to project ([398e901](https://github.com/tatupesonen/formatbot/commit/398e9010ffad8c626e089830d284068ce1af14bf))
+* **ci:** add husky pre-push hook to run tests ([84c53f0](https://github.com/tatupesonen/formatbot/commit/84c53f065b9390179b764e0296c2d35e08b654e5))
+* **ci:** add manual workflow dispatch ([7f18c06](https://github.com/tatupesonen/formatbot/commit/7f18c06178bc1a5767634a5ee0adc78f87b912d2))
+* **core:** add new codeblock parser ([fdc878f](https://github.com/tatupesonen/formatbot/commit/fdc878fb1f362088cad45dd1024abbd119df397e))
+* **core:** add new formatting command ([1559906](https://github.com/tatupesonen/formatbot/commit/1559906d7197717fbf9e4c97e4e1e6ff170463af))
+* show amount of guilds joined in status ([aa0bf7e](https://github.com/tatupesonen/formatbot/commit/aa0bf7e602616184ad5113d1fe610ae0a606ee56))
+
+
+### Bug Fixes
+
+* add IParser to DITypes ([9112520](https://github.com/tatupesonen/formatbot/commit/9112520e210fe69e8786753ce9b9f4994d99c651))
+* **chore:** update lint-staged settings ([077634c](https://github.com/tatupesonen/formatbot/commit/077634c7b9d3b249b95ce010291003612ea57440))
+* **ci:** attempt fixing the yaml of workflow_dispatch ([ea9fae8](https://github.com/tatupesonen/formatbot/commit/ea9fae8264eedf75670f55d285d961094ade828d))
+* **ci:** final fix to the workflow ([acb9031](https://github.com/tatupesonen/formatbot/commit/acb9031707c372678a30cb956483f1e07994bc77))
+* **ci:** sign releases ([b05e31a](https://github.com/tatupesonen/formatbot/commit/b05e31abe9a9481e4103f9c645d0252cb665a7a6))
+* **ci:** try fixing test action ([2e83620](https://github.com/tatupesonen/formatbot/commit/2e836208adbf8540537d4fca3701e12ba3dc5abe))
+* **ci:** try fixing the docker workflow ([3c5d3c2](https://github.com/tatupesonen/formatbot/commit/3c5d3c24161bd77e13041b69809f5745ae01743b))
+* **core:** handle case for no code blocks ([2e01220](https://github.com/tatupesonen/formatbot/commit/2e01220ee798913fe89c3333ba2346325cdfa986))
+* **dependency:** lock black version for now ([bb4b0db](https://github.com/tatupesonen/formatbot/commit/bb4b0dbb4003da1bf35c733ac0e5126f7bfb56b8))
+* remove debug console.logs in format.ts ([35b923f](https://github.com/tatupesonen/formatbot/commit/35b923f6fd592bdd41d9e6690f6c9b9214a07dac))
+* **test:** fix whitespacing in test ([983d27b](https://github.com/tatupesonen/formatbot/commit/983d27b7e9119cb2f903d69fd78a465a9c292b68))
+
 ### [1.3.1](https://github.com/tatupesonen/formatbot/compare/v1.3.0-fix-tensorflow...v1.3.1) (2021-09-02)